trustworthy-reviews/requirements.txt
numpy
pandas
scikit-learn
scipy
regex
unidecode
emoji
langdetect
sentence-transformers
fastapi
uvicorn
streamlit
pyyaml
joblib
tqdm